# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sucursales/{id}/api-key": {
      "post": {
        "description": "La key se devuelve UNA sola vez; en la base queda su hash. Rotar invalida la key anterior de inmediato: el siguiente request del agente con ella es 401. Si llegan dos rotaciones a la vez, las dos responden 201 pero sólo la última key queda vigente.",
        "operationId": "SucursalApiKeyController_rotar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyEmitidaDto"
                }
              }
            }
          },
          "400": {
            "description": "`id` no es un UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente (visor).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La sucursal no existe o es de otra empresa. Misma respuesta en los dos casos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Genera la API key del agente de la sucursal, o la rota si ya tenía una.",
        "tags": [
          "agentes"
        ]
      }
    },
...
    "/empresas": {
      "get": {
        "operationId": "EmpresasController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmpresaDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Empresas en tu alcance: la tuya, o todas si eres admin_global. Activas e inactivas.",
        "tags": [
          "organizacion"
        ]
      },
      "post": {
        "operationId": "EmpresasAdminController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearEmpresaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Da de alta una empresa. Sólo admin_global.",
        "tags": [
          "administracion"
        ]
      }
    },
    "/sucursales": {
      "get": {
        "operationId": "SucursalesController_listar",
        "parameters": [
          {
            "name": "empresaId",
            "required": false,
            "in": "query",
            "description": "Sólo las sucursales de esta empresa. Fuera del alcance del usuario = 404.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SucursalDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "`empresaId` no es un UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La empresa pedida no existe o no está en tu alcance. Misma respuesta (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sucursales en tu alcance, activas e inactivas. Nunca incluye la API key ni su hash.",
        "tags": [
          "organizacion"
        ]
      },
      "post": {
        "operationId": "SucursalesAdminController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearSucursalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucursalDto"
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo. Zona que no es IANA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La empresa: No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Da de alta una sucursal en una empresa de tu alcance. Nace sin API key.",
        "tags": [
          "administracion"
        ]
      }
    },
    "/empresas/{id}": {
      "patch": {
        "operationId": "EmpresasAdminController_editar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarEmpresaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Renombra, da de baja o reactiva una empresa. Sólo admin_global.",
        "tags": [
          "administracion"
        ]
      }
    },
    "/sucursales/{id}": {
      "patch": {
        "description": "Cambiar `zonaHoraria` de una sucursal con historial reclasifica sus ventas pasadas en los cortes por día: los agregados usan la zona actual.",
        "operationId": "SucursalesAdminController_editar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarSucursalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SucursalDto"
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo. Zona que no es IANA.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita nombre y zona de una sucursal, o la da de baja / reactiva.",
        "tags": [
          "administracion"
        ]
      }
    },
    "/usuarios": {
      "get": {
        "operationId": "UsuariosAdminController_listar",
        "parameters": [
          {
            "name": "empresaId",
            "required": false,
            "in": "query",
            "description": "Sólo los usuarios de esta empresa. Fuera de tu alcance = 404.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "`empresaId` no es un UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La empresa pedida: No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Usuarios en tu alcance, activos e inactivos. Nunca incluye el hash.",
        "tags": [
          "administracion"
        ]
      },
      "post": {
        "operationId": "UsuariosAdminController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo. empresaId ausente (o presente para admin_global).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso). O un admin_empresa pidiendo rol admin_global.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La empresa: No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "El email ya está en uso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Da de alta un usuario con rol y empresa.",
        "tags": [
          "administracion"
        ]
      }
    },
    "/usuarios/{id}": {
      "patch": {
        "operationId": "UsuariosAdminController_editar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo. O un cambio de rol no permitido, o darte de baja a ti mismo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso). O un admin_empresa pidiendo rol admin_global.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita nombre y rol de un usuario, o lo da de baja / reactiva.",
        "tags": [
          "administracion"
        ]
      }
    },
    "/usuarios/{id}/password": {
      "post": {
        "description": "Sus refresh tokens dejan de servir en ese instante; su access token vigente, hasta que venza (15 min).",
        "operationId": "UsuariosAdminController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Contraseña restablecida."
          },
          "400": {
            "description": "Cuerpo inválido, `id` que no es UUID, o un PATCH sin ningún campo. O es tu propia contraseña: ésa se cambia en /cuenta/password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token, token inválido o vencido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para la ruta (depende de la ruta, no del recurso).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe o es de otra empresa. Misma respuesta en los dos casos (nunca 403).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restablece la contraseña de un usuario de tu alcance.",
        "tags": [
          "administracion"
        ]
      }
    }
  },
  "info": {
    "title": "Monitor SoftRestaurant API",
    "description": "Contrato de la API del monitor. Fuente única: el frontend construye contra este documento. Lo genera `npm run openapi` en /api a partir de los decoradores.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "monitor_refresh"
      },
      "agente": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Credenciales inválidas"
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ApiKeyEmitidaDto": {
        "type": "object",
        "properties": {
          "sucursalId": {
            "type": "string",
            "format": "uuid"
          },
          "apiKey": {
            "type": "string",
            "example": "msr_3q2-7wXz...",
            "description": "La API key en claro. Se muestra SÓLO en esta respuesta: en la base queda su hash. El agente la manda en el header `X-Api-Key`."
          }
        },
        "required": [
          "sucursalId",
          "apiKey"
        ]
      },
...
      "EmpresaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "nombre",
          "activo"
        ]
      },
...
      "CrearEmpresaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "nombre"
        ]
      },
      "EditarEmpresaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "activo": {
            "type": "boolean",
            "description": "`false` da de baja la empresa: sus usuarios ya no pueden entrar ni refrescar y los agentes de sus sucursales reciben 401. Sus sucursales y usuarios no se tocan."
          }
        }
      },
...
    }
  }
}